# Conflicts:
#	Depo/Depo/App/Depo-AppStore-Info.plist
#	Depo/Depo/App/Depo-Enterprise-Info.plist

+++ Depo/Depo/Presentation/UserStories/Others/AlertFilesActionsSheet/Presenter/AlertFilesActionsSheetPresenter.swift
@@ -108,7 +108,9 @@ class AlertFilesActionsSheetPresenter: MoreFilesActionsPresenter, AlertFilesActi
             var types: [ElementTypes] = [.info, .share, .move]
             
             types.append(item.favorites ? .removeFromFavorites : .addToFavorites)
-            types.append(.moveToTrash)
+            if !item.isReadOnlyFolder {
+                types.append(.moveToTrash)
+            }
             
             if item.fileType == .image || item.fileType == .video {
                 types.append(.download)
@@ -220,6 +222,10 @@ class AlertFilesActionsSheetPresenter: MoreFilesActionsPresenter, AlertFilesActi
                     filteredActionTypes.remove(at: removeFromFavorites)
                 }
                 
+                if remoteItems.first(where: { !$0.isReadOnlyFolder }) == nil {
+                    filteredActionTypes.remove(.moveToTrash)
+                }
+                
                 if let index = filteredActionTypes.index(of: .deleteDeviceOriginal) {
                     MediaItemOperationsService.shared.getLocalDuplicates(remoteItems: remoteItems, duplicatesCallBack: { [weak self] items in
                         let localDuplicates = items
@@ -251,9 +257,9 @@ class AlertFilesActionsSheetPresenter: MoreFilesActionsPresenter, AlertFilesActi
         
         var filteredTypes = types
         let langCode = Device.locale
-        if langCode != "tr" {
-            filteredTypes = types.filter({ $0 != .print })
-        }
+//        if langCode != "tr" {
+//            filteredTypes = types.filter({ $0 != .print }) //FE-2439 - Removing Print Option for Turkish (TR) language
+//        }
         basePassingPresenter?.getSelectedItems { [weak self] selectedItems in
             //FIXME: this part can actualy be wraped into background thread
             guard let self = self else {
@@ -310,7 +316,15 @@ class AlertFilesActionsSheetPresenter: MoreFilesActionsPresenter, AlertFilesActi
                         }
                     }
                 case .hide:
-                    action = UIAlertAction(title: TextConstants.actionSheetHide, style: .default, handler: { _ in
+                    
+                    var title: String
+                    if currentItems.first?.fileType == .faceImageAlbum(.things) {
+                        title = TextConstants.actionSheetHideSingleAlbum
+                    } else {
+                        title = TextConstants.actionSheetHide
+                    }
+                    
+                    action = UIAlertAction(title: title, style: .default, handler: { _ in
                         AnalyticsService.sendNetmeraEvent(event: NetmeraEvents.Actions.ButtonClick(buttonName: .hide))
                         let allowedNumberLimit = NumericConstants.numberOfSelectedItemsBeforeLimits
                         if currentItems.count <= allowedNumberLimit {
